restructure rawgraph package into graph and fringe package
